Further develop CSS styling, almost in good shape
Feeling quite comfortable with the box model of CSS,
however a few things still bother me. The main list in
the home page is not centered, and I want to find out a way
to evenly distribute the recipes in rows or columns.
Tried using span in the text but it is not aligning how
I'm imagining it. Will research on how to achieve it.